+++ changelog.md
@@ -152,3 +152,5 @@ V1.18.0
 * Add vehicle name to pre-existing add overlay
 * Use hard entity pinning
 * Add setting to disable 'add to vehicle' overlay
+* Minor fixes / improvements, especially to descriptions
+* Add warning to deleting an entity