+++ changelog.md
@@ -28,8 +28,9 @@ v3.9.0
 * Supposedly fix jackzvehiclelib improperly doing extras
 * (3.9.1) Updated jackzvehiclelib
 * (3.9.2) Fix extras (Thanks hexarobi)
-* (3.9.3) Actually fix extrasa
-* (3.9.3) Actually fix extrasa
+* (3.9.3) Actually fix extras
+* (3.9.3) Actually fix extras
+* (3.9.4) Fix vehiclelib bug (Thanks hexarobi)
 
 # actions (v1)
 
@@ -121,6 +122,7 @@ V1.16.0
 * (1.16.1) Catch server upload errors
 * (1.16.2) Catch invalid models
 * (1.16.3) Fix false error on successful upload
+* (1.16.4) Fix vehiclelib bug
 
 V1.17.0